Improvements to the benchmarks. (#97)
### What

Improvements to the benchmarks so that:

1. They all work locally without needing to build a local Docker image for the agent.
2. The text summary is always reported.
3. The summary is always published to the historical log.
4. We fail if there are any HTTP errors.

This probably shouldn't be one changeset, but… life is short.

### How

I have rewritten the _local_ scripts to avoid Docker for the agent and k6. CI still uses Docker for these.

This has the nice side effect of changing the benchmarks so they look for an agent on localhost:8100 by default, not agent:8100, which means if you run `k6` yourself, things should just work.

I've also fixed the `handleSummary` function so it also reports the text summary to STDOUT. I thought this was the default behavior if the function returned `undefined` but I guess not.

_select-by-pk.js_ now exports `handleSummary`, and I've moved it to a common file to make it easier. We can move other stuff there too, if we like.

Finally, I've added "thresholds", which ensure that if the checks fail, the run fails.