Expose the scalar types and aggregate functions through the schema. (#140)
### What

We currently support two scalar types: `Int` and `String`.

This specifies the scalar types in the schema. It also adds information to the configuration in order to gather the aggregate functions associated with those types, and adds them to the schema too.

### How

1. I moved the introspection SQL into its own file, so that I can benefit from syntax highlighting.
2. I have added a bunch more SQL to gather the names of aggregate functions associated with integers (`integer`/`smallint`/`bigint`/`numeric`) and strings (`text`).
3. We store this information in the configuration.
4. We expose it through the schema.

To see how what the changes _do_, take a look at the updated schema snapshot and `chinook-deployment.json`.