Pull the models from ndc-sdk (ndc-hub), not ndc-client. (#147)
ndc-sdk re-exports the models. Using them ensures we are working with
the same version of the models expected by the SDK.

This bit me when trying to bump the version of ndc-client to take
advantage of a new piece of functionality without bumping ndc-sdk too.
Better to just not have the option.

It's worth noting that ndc-hub has been renamed to ndc-sdk, but we still
use `ndc_hub` in our codebase for now. We can change this separately.