implement top-level native queries with arguments (#184)
### What

We want to be able to pass arguments to top-level native queries.

### How

1. We parse the native query sql and deconstruct it to textual parts and
parameters when we read the configuration. We also split the metadata.rs
module to database parts and native queries part.
2. We include the arguments when inserting the native queries to the
state.
3. When we build the CTEs, we lookup each query parameter in the
arguments, translate the values we found to sql::ast values, and
generate native query sql that is made of textual parts and parameter
parts. We also moved the value translation code out of filtering.rs and
into a new module values.rs.
4. ???
5. Native queries