Let me just type `just`. (#188)
I don't want to remember all the commands to verify the code is valid, I
just want to `just`.

This adds a new `just check` command as the first one, so it's also the
default. It only delegates to other commands: `build`, `lint`,
`format-check`, and `test`.

It means I can write a one-line Git pre-commit file that just says
`just`.