fix config generation for empty postgres (#196)
### What

The sql we use to generate the configuration of a postgres database can
return `tables` or `aggregate_functions` as null, for example if the
database is empty. We want to make sure the configuration runs correctly
in these cases. See: https://hasurahq.atlassian.net/browse/NDAT-869


### How

Since the sql result parsing code expects a non-null value, we wrap the
field selection in the query with `coalesce`, and return an empty object
when a `null` (no rows) are returned.

I have tested this manually by spinning an empty postgres, removing the
postgis extension, observing the failure, fixing the bug, and observing
that it now runs.