Add `ndc-cockroach` (#167)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

Add a new `ndc-cockroach` data connector reusing most of the
`ndc-postgres` code.

### How

We mostly use functions imported from `ndc-postgres` itself. We use
shared test fixtures, but the tests themselves are run in each NDC
folder.

This does not include changes to introspection code, so those tests are
commented out so that we can unblock the other NDC flavours.