Expose the comparison operators through the schema. (#208)
### What

This adds the comparison operators for each type to the schema.

It does not change the names of the operators to be prefixed with an
underscore. I wanted to do this separately to make the diff easier to
understand.

### How

1. We iterate through scalar types directly, not by way of the aggregate
functions. To make this easier, I added the `enum-iterator` crate.
2. The `BinaryOperator` enum has been duplicated in the
`metadata::database` module, and augmented with behavior so we can get
the operator's name and the type of its RHS argument (currently always
the same as the LHS argument).

---------

Co-authored-by: Gil Mizrahi <gil@hasura.io>