Create ndc-aurora (#219)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

This add an `ndc-aurora` connector for use with AWS Aurora.

### How

This works with all the functionality of Postgres for now (this may
change in future though).

Testing and local development are a bit different. Instead of spinning
up stuff in Docker, we use a static AWS instance for local/CI testing.
To run `just dev-aurora` you'll need to populate the
`AURORA_CONNECTION_STRING` env var from details in 1Password.

We then smash that env var into the deployment before running local dev,
and in CI. Hence, the generated `static/aurora/chinook-deployment.json`
is gitignored so we don't check the credentials into git.

### What is missing

Introspection does not work against a read-only replica. This is true
for all Postgres flavours, and there is now a ticket for fixing this by
avoiding the use of temporary functions.