Reformat everything with Prettier. (#222)
### What

We seem to be going back and forth on formatting, especially for
JavaScript. Let's stop doing that and just use an auto-formatter for
JavaScript, JSON, YAML, and Markdown.

This replaces the `jq`-based formatting for the goldenfiles; the results
seem to be identical.

We ignore `flake.lock` because it is machine-generated and
auto-formatting it differently from `nix flake update` will cause
unnecessary grief.

### How

`just format` now runs `prettier --write .`, which reformats JavaScript,
JSON, YAML, and Markdown, ignoring all files in `.gitignore` or
`.prettierignore`. I have added `flake.lock` to `.prettierignore` to
ensure we don't thrash between two formats.

`just format-check` will now run `prettier --check .`.

I have run `just format` to ensure that everything is formatted
consistently. We can change this by configuring Prettier with a
`.prettierrc` file if we don't like the current formatting.