wrap a few Strings in newtypes (#237)
### What

We give newtypes to a few components which were previously simple
`String`s, rename a couple of types, and add some docstrings on types in
sql::ast. If this is too much change in one PR let me know and I'll
split it to two.

### How

- The most important change in this PR is in the sql::ast module. We
rename `TableName` and `ColumnName` to `TableReference` and
`ColumnReference`, and add three other newtypes - `ColumnName`,
`TableName`, and `SchemaName` around what were previously simple
`String`s.
- The second important change is making `collection.lookup_column`
return a custom `ColumnInfo` that instead of a `String` has a
`ColumnName`
- We also rename `Expression:[Arity]Operator` to
`Expression::[Arity]Operation`.
- The rest of the changes are just following the errors and correcting
the types.

### Why

As a consequence of our change, in order to build a `TableColumn`, we
now need to provide a `ColumnName` instead of `String`.
[Previously](https://github.com/hasura/postgres-ndc/pull/228) we would
sometimes use the `String` we get from `QueryRequest` representing the
name of the string in the table metadata, and not the actual column name
in the table which we need to lookup using `collection.lookup_column`.

```diff
-    TableColumn { table: TableName, name: String },
+    TableColumn {
+        table: TableReference,
+        name: ColumnName,
+    },
```

This change makes it easier to do the right thing (use the column name
after lookup when querying the table), and harder to do the wrong thing
(use the column alias we get from `QueryRequest`).