use native postgres types instead of custom types (#245)
### What

https://hasurahq.atlassian.net/browse/NDAT-894

We want our connector to expose postgres types instead of custom types
such as `String`, `Int`, etc.

### How

I'm sorry for the gigantic PR. The actual code change is very simple -
we change the configuration.sql to generate postgres types instead of
giving them custom names and changing the `ScalarType` enum. The rest of
the PR is generating the deployments files again and renaming the types
in the query-engine tests.

Maybe better reviewed as a sequence of commits rather than all at once.