use #[serde(default)] and output jsonschema (#249)
This PR tracks the JSONSchema for the deployment configuration in the
repo and adds a test to ensure it's up to date. It also changes `schemars` to preserve the order of the fields as defined in Rust in the JSONSchema definitions.

It also makes the `metadata` and `aggregate_functions` fields in the
configuration optional since they have default values.