chore: remove nix-build from nix check job (#250)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

Our Nix checks currently build every binary with Nix in series. This can
take [~30
minutes](https://github.com/hasura/postgres-ndc/actions/runs/6142614100/job/16664640226),
and is somewhat duplicated effort, so removing it.

### How

Removed the Github Action job. We already test that Cargo builds, and
that `nix develop` and `nix flake check` work, so this is overkill. If
we really must bring it back, let's at least split it up so it builds
each binary in parallel.