chore: fix cockroach introspection (#214)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

Makes a new `configuration.rs` and `configuration.sql` for
`ndc-cockroach` rather than using the `postgres` one.

### How

* We cannot have user-defined functions, so we have to inline (and if we
were to zoom forwards to returning unprocessed type names then we
wouldn't need this at all)
* We cannot safely do `json_object_agg` in a lateral join or correlated
sub-select because of
https://github.com/cockroachdb/cockroach/issues/109856.
* The reason we end up in the poorly handled null-case of
`json_object_agg` in the first place is that
`information_schema.table_constraint_usage` (IIRC) is just empty.
https://hasurahq.atlassian.net/browse/NDAT-887 is meant to address this
* So we have to do left-outer-join+group-by to produce the same effect
as the sub-select. This is probably the main independent change of this
PR.

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->

---------

Co-authored-by: Philip Lykke Carlsen <plcplc@gmail.com>