implement native queries handling in missing branches (#254)
### What

There were a few instances where we did not implement handling for
native queries in the code. Specifically, these were in filtering and
sorting by a relationship collection that was a native query. This PR
implements these cases.

### How

Instead of guarding against native queries and building a `FROM` clause
manually, we grab the 'arguments' field and build a `FROM` clause from
the relevant function in `root` which is responsible for it.

We also added a couple of tests that check that native queries can be
used in the scenarios mentioned above using the existing native queries
we have defined in the configuration.