Make it possible to add health checks to the Docker container. (#265)
### What

We need to be able to specify a health check for the Docker container.
This makes it possible and adds documentation.

### How

First, I upgraded NDC SDK to the latest version, which adds a
`check-health` command to the binary. This means we can run e.g.
`ndc-postgres check-health`.

Secondly, I patched the Docker image to ensure that there is a symlink
to `ndc-postgres` (or equivalent, for other flavors) in `/bin`, so it's
on the `PATH`.

Thirdly, I documented how to run the connector with Docker Compose,
including a `healthcheck` section.