infer comparison target types (#263)
### What

The user may pass literal arguments to ndc-postgres. We need to know
what the type of the values are so we can pass them as prepared
arguments via sqlx. This PR infers the type of values from the context.

### How

1. For native queries arguments, the type of arguments is defined in the
native query definition, we just need to fetch it.
2. For comparisons, we can infer the type of a value from the operator
and the column. We simply need to fetch the type from the environment.
3. We pattern match on the json value we receive and the scalar type we
inferred and either convert it to built-in types (`i64`, `f64`,
`String`), or we pass the as `cast('<string>' to <type>)`.