A bunch of changes to make running tests more reliable. (#267)
### What

This fixes a couple of `just` commands so they work as expected.

`just test` no longer tries to run tests that are guaranteed to fail.

`just generate-chinook-configuration` no longer results in an empty file
if part of the pipeline fails.

### How

`just test` was failing for me on AWS Aurora tests because I don't have
an Aurora connection string. I have put the Aurora tests behind a Cargo
feature, disabled by default, and enabled it if the
`AURORA_CONNECTION_STRING` environment variable is set and non-empty.

`just generate-chinook-configuration` was generating empty
deployment files for me. I've fixed the script so it fails properly and
leaves the deployment file alone if anything goes wrong.