Refactor metrics to make them easier to extend. (#272)
### What

These are the changes I made in order to make it easier for me to add
new metrics and ensure I don't break existing ones.

It encapsulates more metrics-related logic, and removes duplication.

### How

- Pool options metrics are now only recorded at startup, because they do
not change.
- Some functions have been moved into `impl Metrics` because method
syntax can be easier to follow. (Call it "idiomatic Rust" if you must.)
- We now use the aliases for metrics types because they're simpler (e.g.
`IntGauge` instead of `GenericGauge<AtomicI64>`).
- Metrics are private to the module, and can only be manipulated through
methods on `Metrics`. This means that methods have been added to
increment the counters.
- Some duplicate code has been factored out into functions.