Record query plan and execution times with a timer. (#274)
### What

This adds two new metrics: `query_plan_time` and `query_execution_time`,
which are histograms tracking counts, bucketed roughly exponentially by
time taken in seconds.

We're using the default buckets, which are, [as of `prometheus`
v0.13.3](https://docs.rs/prometheus/0.13.3/src/prometheus/histogram.rs.html#25-27):
`0.005, 0.01, 0.025, 0.05, 0.1, 0.25, 0.5, 1.0, 2.5, 5.0, 10.0`, with an
implicit `+Inf` bucket. Each bucket is "less than or equal", which means
that, for example, if a request takes 0.8s, it will increment the
counters for each of the the `1.0`, `2.5`, `5.0`, `10.0`, and `+Inf`
buckets.

We are currently only tracking successful queries; the timing values for
failed queries are discarded.

### How

First of all, I pulled out functions from the `query` function.

I then added the histograms to the metrics, and exposed `time_<metrics>`
methods. These return a wrapped `Timer` which can only be completed with
a `Result<T, E>`. If the result is `Ok`, the data is stored; if it's an
`Err`, it's discarded. I wrapped it to avoid running into problems in
the future, as this seemed quite error-prone.

I tried a different abstraction first, exposing a `time_<metric>`
function that took an `FnOnce() -> Result<T, E>` and handled all the
details internally. Unfortunately, because it really needs to return a
future, the signature eventually ended up like this:

```
pub async fn time_metric<T, E, F>(action: impl FnOnce() -> F) -> Result<T, E>
where
    F: std::future::Future<Output = Result<T, E>>
```

Frankly, I hated it. The real showstopper, though, was that async
closures are gated behind a nightly-only flag.