chore: query column nullability in introspection (#302)
<!-- The PR description should answer 2 (maybe 3) important questions:
-->

### What

Previously we did not query column nullability in introspection, and
somewhere decided that meant everything was non-nullable.

### How

Queries `introspection_schema.columns.is_nullable` to find out which
columns are nullable and returns it in introspection. We default to
`Nullable` if we do not receive a value, so old deployments are still
valid.