Pass-through scalar types (#311)
### What

This PR changes how scalar types are handled. Rather than explicitly
declaring which types we know about, we consider them opaque values that
we pass along.

### Why

The main advantage to being hands-off about scalar types is that all
types receive a uniform treatment. There is no room for us to
accidentally introduce behavior that will require breaking changes to
correct.

Users benefit from the ndc layer being as transparent and predictable as
possible. They are free to make any customizations at the DDN layer or
(if they are privileged) in the database itself.

This also simplifies our implementation and allows us to delete a lot of
code that was essentially just mapping cases 1-to-1.

### How

All the changes essentially fall out from changing `enum ScalarType =
..` to `struct ScalarType(String)`.

There are a few noteworthy wrinkles to this:

* Everyplace a value enters the SQL query from a json value becomes a
Cast.
* Casts were mischaracterised as `Value`s, but it's actually an
`Expression`.
* There are some scalar types we might still want to censor, since they
seem to require a better introspection query (see the chinook deployment
file). Those are in particular `ARRAY`, `anyelement` etc, and whatever
it is that turns up as `USERDEFINED`.
*

---------

Co-authored-by: Gil Mizrahi <gil@hasura.io>